+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.0.0-dev.3
 **💥 Breaking Changes**\
 * Refactored `AlarmSettings` model.
+* Ignore oncoming alarm if another alarm is ringing.
 * [Android] Update kotlin version to `1.8.0`.
 
 ## 4.0.0-dev.2